Streamline UI tab into General/Editor/Pipeline Colors; inline reset buttons
Per review:

1. The UI tab's flat control grid is now three named sections:
   General (theme, accordion change alerts, multi-step size, execution
   delay), Editor (vi mode, font size), Pipeline Colors (unchanged).
   Split the old UiSettingsPanel/ExecutionSettingsPanel into three
   focused components — GeneralSettingsPanel, EditorSettingsPanel,
   PipelineColorsPanel — matching the CPU/Cache split already used on
   the Simulation tab.

2. Both "Reset to defaults" buttons move out of the shared
   DialogActions bar and into the bottom of their own tab's content:
   "Reset UI to defaults" at the end of the UI tab, "Reset Simulation
   to defaults" at the end of Simulation. DialogActions now only has
   Close.

Since the Simulation tab's content — including its own reset button —
only renders while that tab is active, and the tab itself is
unselectable while a program is running, "Reset Simulation to
defaults" is no longer just disabled while running: it's unreachable,
same as CPU Forwarding and the cache fields. Updated the
"disabled while running" coverage in settings-dialog.spec.js to match
(toHaveCount(0) rather than toBeDisabled() while parked on the UI
tab).

Verified: tsc --noEmit, eslint (one preexisting-pattern warning, no
errors), vitest (200 passed), and the full touched-spec set + control
(53 passed, 1 pre-existing unrelated skip).
